requirements.txt
langchain==0.3.18
langchain-google-genai==2.0.8
langchain-community==0.3.16
faiss-cpu==1.13.1
pandas==2.2.3
python-dotenv==1.0.1
google-generativeai==0.8.3
sentence-transformers==3.3.1
